Ye_Ethiopia_lij_back/Ye_Ethiopia_lij/requirements.txt
asgiref==3.11.1
channels==4.2.2
channels-redis==4.2.1
daphne==4.1.2
attrs==25.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-mongodb-backend==6.0.2
django-filter==24.1
django-dotenv==1.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
PyJWT==2.11.0
pymongo==4.16.0
PyYAML==6.0.3
python-dotenv==1.0.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
